# src/backend/CompanyHero.Api/Contracts/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "CompanyHero API",
    "description": "Verbindlicher Vertrag nach A-009 und K13: Kennungen und Dezimalwerte als Strings, Zeitpunkte mit Offset, Null und fehlendes Feld unterscheidbar.",
    "version": "1"
  },
  "paths": {
...
    "/api/me": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "GetMe",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/join/{code}": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "PreviewJoin",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinPreviewResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "Join",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/join/{code}/passkey-options": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "BeginJoinPasskey",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinPasskeyOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyCeremonyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/join/{code}/kiosk": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "JoinAtKiosk",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KioskJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/join/role/{code}": {
      "get": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "PreviewRoleJoin",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinPreviewResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "JoinWithRoleCode",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/join/role/{code}/passkey-options": {
      "post": {
        "tags": [
          "CompanyHero.Api"
        ],
        "operationId": "BeginRoleJoinPasskey",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinPasskeyOptionsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasskeyCeremonyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JoinPasskeyOptionsRequest": {
        "required": [
          "displayName"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          }
        }
      },
      "JoinPreviewResponse": {
        "required": [
          "tenantId",
          "tenantName",
          "ways",
          "pendingExternal",
          "role"
        ],
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "tenantName": {
            "type": "string"
          },
          "ways": {
            "$ref": "#/components/schemas/JoinWaysResponse"
          },
          "pendingExternal": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PendingExternalResponse"
              }
            ]
          },
          "role": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "JoinRequest": {
        "required": [
          "displayName",
          "visibility",
          "passkey",
          "email",
          "useExternal",
          "kioskPin"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/VisibilityDto"
          },
          "passkey": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PasskeyAnswerRequest"
              }
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "useExternal": {
            "type": "boolean"
          },
          "kioskPin": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "JoinResponse": {
        "required": [
          "tenantId",
          "personId",
          "kioskId",
          "recoveryCode",
          "sessionKind"
        ],
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string"
          },
          "personId": {
            "type": "string"
          },
          "kioskId": {
            "type": "string"
          },
          "recoveryCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "sessionKind": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SessionKindDto"
              }
            ]
          }
        }
      },
      "JoinWaysResponse": {
        "required": [
          "passkey",
          "magicLink",
          "kiosk",
          "providers",
          "forcedProviderKeys"
        ],
        "type": "object",
        "properties": {
          "passkey": {
            "type": "boolean"
          },
          "magicLink": {
            "type": "boolean"
          },
          "kiosk": {
            "type": "boolean"
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderResponse"
            }
          },
          "forcedProviderKeys": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "JsonElement": {},
...
      "KioskJoinRequest": {
        "required": [
          "displayName",
          "visibility",
          "pin"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/VisibilityDto"
          },
          "pin": {
            "type": "string"
          }
        }
      },
...
      "MeResponse": {
        "required": [
          "personId",
          "displayName",
          "tenantId",
          "tenantName",
          "roles"
        ],
        "type": "object",
        "properties": {
          "personId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "tenantName": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PasskeyAnswerRequest": {
        "required": [
          "state",
          "credential",
          "deviceName"
        ],
        "type": "object",
        "properties": {
          "state": {
            "type": "string"
          },
          "credential": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "deviceName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "PasskeyCeremonyResponse": {
        "required": [
          "options",
          "state"
        ],
        "type": "object",
        "properties": {
          "options": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "state": {
            "type": "string"
          }
        }
      },
...
      "PendingExternalResponse": {
        "required": [
          "providerKey",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "providerKey": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProviderResponse": {
        "required": [
          "key",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
...
      "RoleJoinRequest": {
        "required": [
          "realName",
          "visibility",
          "passkey",
          "email",
          "useExternal"
        ],
        "type": "object",
        "properties": {
          "realName": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/VisibilityDto"
          },
          "passkey": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PasskeyAnswerRequest"
              }
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "useExternal": {
            "type": "boolean"
          }
        }
      },
...
      "SessionKindDto": {
        "enum": [
          "member",
          "privileged",
          "kiosk_device",
          "kiosk_person"
        ]
      },
...
      "VisibilityDto": {
        "enum": [
          "only_me",
          "team",
          "company"
        ]
      },
...
    }
  },
...
}